# (fix) - O3-3684 - improve performance of age() function