# Requirements

- [x] This PR has a title that briefly describes the work done including the ticket number. Ensure your PR title includes a [conventional commit](https://o3-docs.openmrs.org/docs/frontend-modules/contributing#contributing-guidelines) label (such as `feat`, `fix`, or `chore`, among others). See existing PR titles for inspiration.

## For changes to apps

- [ ] My work conforms to the [**O3 Styleguide**](https://om.rs/styleguide) and [**design documentation**](https://om.rs/o3ui).

## If applicable

- [ ] My work includes tests or is validated by existing tests.
- [ ] I have updated the [esm-framework mock](https://github.com/openmrs/openmrs-esm-core/blob/main/packages/framework/esm-framework/mock.tsx) to reflect any API changes I have made.

## Summary
<!-- Please describe what problems your PR addresses. -->
The `age()` function refactored in this [PR](https://github.com/openmrs/openmrs-esm-core/pull/1100) uses a [`DurationFormat`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DurationFormat) polyfill that is a really expensive object to create (takes 66 - 100ms on my laptop depending on whether it's plugged in). We rely on it in the `esm-ward-app` to render the age, time since admission and time on ward for each patient card. That's about 200ms per patient on those 3 elements. On a ward with ~5 patients, the rendering is cripplingly slow.

This PR just adds caching to the `DurationFormat` object so we only get the performance penalty on the first `age()` call.

## Screenshots
<!-- Required if you are making UI changes. -->
Profiling before this change. Those 3 elements take 200ms to render:
![image](https://github.com/user-attachments/assets/2b0861f3-9913-4e89-b106-6566a9f0797a)

Profiling after this change. PatientAge element takes 100ms on first render, but is reasonable faster afterwards. The other 2 time elements aren't even showing in the waterfall (probably because their render times are insignificant).
![withformatjs](https://github.com/user-attachments/assets/5404dc2e-fb5d-49b6-aa3e-20a205a3d057)


## Related Issue
<!-- Paste the link to the Jira ticket here if one exists. -->
<!-- https://issues.openmrs.org/browse/O3- -->

## Other
<!-- Anything not covered above -->
I initially tried a solution that does away with the polyfill altogether. It eliminates the first call penalty, but is a bit iffy with the i18n stuff. I checked it in [here](https://github.com/openmrs/openmrs-esm-core/commit/2b531bde02bd556bb45cd3bd7c4986be80e4c6e5) in case we want to not use the polyfill.