@ibacher 2024-08-17
Two comments: 

* We should probably cache by locale to support locale switching (specifically in the case where a user has a locale preference that’s different from the default locale for an implementation).
* Reading through some of the issues, it seems like the default locale matching algorithm (“best fit”—which is unspecified, but formatjs tries to use a the Unicode LDML algorithm which is comprehensive and aims to be correct, but slow). It may help substantially cut down the time by switching to the “lookup” algorithm via the `localeMatcher` option to the constructor. While less comprehensive, this should be sufficient for our purposes. This would likely remove much of the remaining overhead. See [the discussion here](https://github.com/formatjs/formatjs/issues/4463).