@chibongho 2024-08-18
`localeMatcher: "lookup"` worked beautifully. < 1ms of runtime. No caching needed.